// web/projects/portal/src/app/vsobjects/model/layout/guide-bounds.ts
/*
 * inetsoft-web - StyleBI is a business intelligence web application.
 * Copyright © 2024 InetSoft Technology (info@inetsoft.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affrero General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
export enum GuideBounds {
   GUIDES_NONE = 0,
   GUIDES_16_9_PORTRAIT = 1,
   GUIDES_4_3_PORTRAIT = 2,
   GUIDES_16_9_LANDSCAPE = 3,
   GUIDES_4_3_LANDSCAPE = 4
}